requirements.txt
flask==2.3.3
flask-session==0.5.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
google-generativeai==0.3.0
python-dotenv==1.0.0
dateparser==1.1.8
unicodedata2==15.1.0
pickle-mixin==1.0.2